The ToPoints and so on where already implemented, so some cleaning
From the PR, we kept also the row renderer, so a collection of Row is also renderable easily using the Charts

From the PR, the simpler Sampler has also be kept in order to sample data for charts.

notebooks have been cleaned, the DataFrame example has been extended to show how to deal with DF and Pivot.
