+++ notebooks/cassandra/case-class-error.snb
@@ -13,6 +13,7 @@
     "customRepos" : null,
     "customDeps" : [ "com.datastax.spark % spark-cassandra-connector-java_2.10 % 1.3.0-M1" ],
     "customImports" : [ "import java.util.UUID", "import scala.util._", "import org.apache.spark.SparkContext._", "import com.datastax.spark.connector._" ],
+    "customArgs" : null,
     "customSparkConf" : {
       "spark.cassandra.connection.host" : "172.17.0.2"
     }
